Used JsonVIEW to remove duplicate code
IBP-3212
