+++ src/main/java/org/ibp/api/brapi/v1/study/StudyDetailsData.java
@@ -2,7 +2,11 @@ package org.ibp.api.brapi.v1.study;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+import com.fasterxml.jackson.annotation.JsonView;
+import org.generationcp.middleware.service.api.BrapiView;
 import org.ibp.api.brapi.v1.location.Location;
+import org.ibp.api.brapi.v2.study.EnvironmentParameter;
+import org.ibp.api.brapi.v2.study.ExperimentalDesign;
 import org.pojomatic.Pojomatic;
 import org.pojomatic.annotations.AutoProperty;
 
@@ -13,15 +17,25 @@ import java.util.Map;
 
 @AutoProperty
 @JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonPropertyOrder({"studyDbId", "studyName", "studyType", "seasons", "trialDbId", "trialName", "startDate", "endDate", "active",
-	"location", "contacts", "additionalInfo"})
 public class StudyDetailsData {
 
 	private Integer studyDbId;
 
 	private String studyName;
 
+	@JsonView(BrapiView.BrapiV2.class)
+	private String studyDescription;
+
 	private String studyType;
+
+	@JsonView(BrapiView.BrapiV2.class)
+	private String lastUpdate;
+
+	@JsonView(BrapiView.BrapiV2.class)
+	private String commonCropName;
+
+	@JsonView(BrapiView.BrapiV2.class)
+	private String studyPUI;
 
 	private List<String> seasons;
 
@@ -36,6 +50,33 @@ public class StudyDetailsData {
 	private Boolean active;
 
 	private Location location;
+
+	@JsonView(BrapiView.BrapiV2.class)
+	private String culturalPractices;
+
+	@JsonView(BrapiView.BrapiV2.class)
+	private List<String> dataLinks;
+
+	@JsonView(BrapiView.BrapiV2.class)
+	private String documentationURL;
+
+	@JsonView(BrapiView.BrapiV2.class)
+	private List<EnvironmentParameter> environmentParameters;
+
+	@JsonView(BrapiView.BrapiV2.class)
+	private ExperimentalDesign experimentalDesign;
+
+	@JsonView(BrapiView.BrapiV2.class)
+	private List<String> externalReferences;
+
+	@JsonView(BrapiView.BrapiV2.class)
+	private String growthFacility;
+
+	@JsonView(BrapiView.BrapiV2.class)
+	private String license;
+
+	@JsonView(BrapiView.BrapiV2.class)
+	private String observationUnitsDescription;
 
 	private List<Contact> contacts = new ArrayList<>();
 
@@ -63,12 +104,17 @@ public class StudyDetailsData {
 	 * @param contacts
 	 * @param additionalInfo
 	 */
-	public StudyDetailsData(final Integer studyDbId, final String studyName, final String studyType, final List<String> seasons,
-			final Integer trialDbId, final String trialName, final String startDate, final String endDate, final Boolean active,
-			final Location location, final List<Contact> contacts, final Map<String, String> additionalInfo) {
+	public StudyDetailsData(final Integer studyDbId, final String studyName, final String studyDescription, final String studyType, final String studyPUI,
+		final List<String> seasons,	final Integer trialDbId, final String trialName, final String startDate, final String endDate,
+		final Boolean active, final Location location, final String culturalPractices, final List<String> dataLinks,
+		final String documentationURL, final List<EnvironmentParameter> environmentParameters, final ExperimentalDesign experimentalDesign,
+		final List<String> externalReferences, final String growthFacility, final String lastUpdate, final String license,
+		final String observationUnitsDescription, final List<Contact> contacts, final Map<String, String> additionalInfo) {
 		this.studyDbId = studyDbId;
 		this.studyName = studyName;
+		this.studyDescription = studyDescription;
 		this.studyType = studyType;
+		this.studyPUI = studyPUI;
 		this.seasons = seasons;
 		this.trialDbId = trialDbId;
 		this.trialName = trialName;
@@ -76,6 +122,16 @@ public class StudyDetailsData {
 		this.endDate = endDate;
 		this.active = active;
 		this.location = location;
+		this.culturalPractices = culturalPractices;
+		this.dataLinks = dataLinks;
+		this.documentationURL = documentationURL;
+		this.environmentParameters = environmentParameters;
+		this.experimentalDesign = experimentalDesign;
+		this.externalReferences = externalReferences;
+		this.growthFacility = growthFacility;
+		this.lastUpdate = lastUpdate;
+		this.license = license;
+		this.observationUnitsDescription = observationUnitsDescription;
 		this.contacts = contacts;
 		this.additionalInfo = additionalInfo;
 	}
@@ -125,6 +181,22 @@ public class StudyDetailsData {
 	 */
 	public StudyDetailsData setStudyType(final String studyType) {
 		this.studyType = studyType;
+		return this;
+	}
+
+	/**
+	 * @return the Study PUI
+	 */
+	public String getStudyPUI() {
+		return studyPUI;
+	}
+
+	/**
+	 * @param studyPUI
+	 * @return StudyDetailsData
+	 */
+	public StudyDetailsData setStudyPUI(final String studyPUI) {
+		this.studyPUI = studyPUI;
 		return this;
 	}
 
@@ -269,6 +341,114 @@ public class StudyDetailsData {
 	 */
 	public StudyDetailsData setTrialName(final String trialName) {
 		this.trialName = trialName;
+		return this;
+	}
+
+	public String getStudyDescription() {
+		return studyDescription;
+	}
+
+	public StudyDetailsData setStudyDescription(final String studyDescription) {
+		this.studyDescription = studyDescription;
+		return this;
+	}
+
+	public String getCulturalPractices() {
+		return culturalPractices;
+	}
+
+	public StudyDetailsData setCulturalPractices(final String culturalPractices) {
+		this.culturalPractices = culturalPractices;
+		return this;
+	}
+
+	public List<String> getDataLinks() {
+		return dataLinks;
+	}
+
+	public StudyDetailsData setDataLinks(final List<String> dataLinks) {
+		this.dataLinks = dataLinks;
+		return this;
+	}
+
+	public String getDocumentationURL() {
+		return documentationURL;
+	}
+
+	public StudyDetailsData setDocumentationURL(final String documentationURL) {
+		this.documentationURL = documentationURL;
+		return this;
+	}
+
+	public List<EnvironmentParameter> getEnvironmentParameters() {
+		return environmentParameters;
+	}
+
+	public StudyDetailsData setEnvironmentParameters(final List<EnvironmentParameter> environmentParameters) {
+		this.environmentParameters = environmentParameters;
+		return this;
+	}
+
+	public ExperimentalDesign getExperimentalDesign() {
+		return experimentalDesign;
+	}
+
+	public StudyDetailsData setExperimentalDesign(final ExperimentalDesign experimentalDesign) {
+		this.experimentalDesign = experimentalDesign;
+		return this;
+	}
+
+	public List<String> getExternalReferences() {
+		return externalReferences;
+	}
+
+	public StudyDetailsData setExternalReferences(final List<String> externalReferences) {
+		this.externalReferences = externalReferences;
+		return this;
+	}
+
+	public String getGrowthFacility() {
+		return growthFacility;
+	}
+
+	public StudyDetailsData setGrowthFacility(final String growthFacility) {
+		this.growthFacility = growthFacility;
+		return this;
+	}
+
+	public String getLastUpdate() {
+		return lastUpdate;
+	}
+
+	public StudyDetailsData setLastUpdate(final String lastUpdate) {
+		this.lastUpdate = lastUpdate;
+		return this;
+	}
+
+	public String getCommonCropName() {
+		return commonCropName;
+	}
+
+	public StudyDetailsData setCommonCropName(final String commonCropName) {
+		this.commonCropName = commonCropName;
+		return this;
+	}
+
+	public String getLicense() {
+		return license;
+	}
+
+	public StudyDetailsData setLicense(final String license) {
+		this.license = license;
+		return this;
+	}
+
+	public String getObservationUnitsDescription() {
+		return observationUnitsDescription;
+	}
+
+	public StudyDetailsData setObservationUnitsDescription(final String observationUnitsDescription) {
+		this.observationUnitsDescription = observationUnitsDescription;
 		return this;
 	}
