+++ src/main/java/org/ibp/api/brapi/v1/study/StudyMapper.java
@@ -1,7 +1,10 @@
 package org.ibp.api.brapi.v1.study;
 
+import org.generationcp.middleware.domain.etl.MeasurementVariable;
 import org.generationcp.middleware.service.api.study.StudyDetailsDto;
 import org.generationcp.middleware.service.api.user.UserDto;
+import org.ibp.api.brapi.v2.study.EnvironmentParameter;
+import org.ibp.api.brapi.v2.study.ExperimentalDesign;
 import org.ibp.api.mapper.ApiMapper;
 import org.modelmapper.Converter;
 import org.modelmapper.ModelMapper;
@@ -39,6 +42,33 @@ public class StudyMapper {
 
 	}
 
+	private static class EnvironmentParameterConverter implements Converter<List<MeasurementVariable>, List<EnvironmentParameter>> {
+
+		@Override public List<EnvironmentParameter> convert(final MappingContext<List<MeasurementVariable>, List<EnvironmentParameter>> context) {
+			final List<EnvironmentParameter> environmentParameters = new ArrayList<>();
+			for (final MeasurementVariable variable : context.getSource()) {
+				final EnvironmentParameter environmentParameter = new EnvironmentParameter();
+				environmentParameter.setDescription(variable.getDescription());
+				environmentParameter.setValue(variable.getValue());
+				environmentParameter.setUnit(variable.getScale());
+				environmentParameter.setParameterName(variable.getName());
+				environmentParameters.add(environmentParameter);
+			}
+			return context.getMappingEngine().map(context.create(environmentParameters, context.getDestinationType()));
+		}
+
+	}
+
+	private static class ExperimentalDesignConverted implements Converter<String, ExperimentalDesign> {
+
+		@Override public ExperimentalDesign convert(final MappingContext<String, ExperimentalDesign> context) {
+			final ExperimentalDesign experimentalDesign = new ExperimentalDesign();
+			experimentalDesign.setDescription(context.getSource());
+			return context.getMappingEngine().map(context.create(experimentalDesign, context.getDestinationType()));
+		}
+
+	}
+
 	private static void addStudyDetailsDataMapping(final ModelMapper mapper) {
 		mapper.addMappings(new PropertyMap<StudyDetailsDto, StudyDetailsData>() {
 
@@ -53,8 +83,11 @@ public class StudyMapper {
 				this.map().setStudyType(this.source.getMetadata().getStudyType());
 				this.map().setTrialName(this.source.getMetadata().getTrialName());
 				this.map().setTrialDbId(this.source.getMetadata().getTrialDbId());
+				this.map().setStudyDescription(this.source.getMetadata().getStudyDescription());
+				this.map().setLastUpdate(this.source.getMetadata().getLastUpdate());
+				this.using(new ExperimentalDesignConverted()).map(this.source.getMetadata().getExperimentalDesign()).setExperimentalDesign(null);
 				this.using(new ContactConverter()).map(this.source.getContacts()).setContacts(null);
-
+				this.using(new EnvironmentParameterConverter()).map(this.source.getEnvironmentParameters()).setEnvironmentParameters(null);
 			}
 		});
 	}
