+++ src/main/java/org/ibp/api/brapi/v2/study/StudyResourceV2Brapi.java
@@ -10,8 +10,11 @@ import org.generationcp.middleware.service.api.location.LocationFilters;
 import org.generationcp.middleware.service.api.study.StudyDetailsDto;
 import org.ibp.api.brapi.v1.common.Metadata;
 import org.ibp.api.brapi.v1.common.Pagination;
-import org.ibp.api.brapi.v2.location.Location;
-import org.ibp.api.brapi.v2.location.LocationMapper;
+import org.ibp.api.brapi.v1.location.Location;
+import org.ibp.api.brapi.v1.location.LocationMapper;
+import org.ibp.api.brapi.v1.study.StudyDetails;
+import org.ibp.api.brapi.v1.study.StudyDetailsData;
+import org.ibp.api.brapi.v1.study.StudyMapper;
 import org.ibp.api.java.study.StudyService;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
