+++ src/main/java/org/ibp/api/java/impl/middleware/study/StudyServiceImpl.java
@@ -314,11 +314,6 @@ public class StudyServiceImpl implements StudyService {
 	}
 
 	@Override
-	public StudyDetailsDto getStudyDetailsForGeolocation(final Integer geolocationId) {
-		return this.middlewareStudyService.getStudyDetailsForGeolocation(geolocationId);
-	}
-
-	@Override
 	public StudyDetailsDto getStudyDetailsByGeolocation(final Integer geolocationId) {
 		return this.middlewareStudyService.getStudyDetailsByGeolocation(geolocationId);
 	}
